+++ arpeggio/__init__.py
@@ -20,7 +20,7 @@ import bisect
 from arpeggio.utils import isstr
 import types
 
-__version__ = "1.9.2"
+__version__ = "1.10.0"
 
 if sys.version < '3':
     text = unicode